requirements.txt
pandas>=2.2
matplotlib>=3.8
seaborn>=0.13
joblib >= 1.3.0
scikit-learn>=1.2,<1.7
xgboost==3.1.2
fastapi>=0.115.12
jinja2>=3.1.6
numpy>=2.2.5
pillow>=11.2.1
pydub>=0.25.1
python-dotenv>=1.1.0
python-multipart>=0.0.20
pyyaml>=6.0.2
requests>=2.32.3
soundfile>=0.13.1
starlette>=0.46.2
timm>=1.0.15
torch>=2.7.0
torchaudio>=2.7.0
torchvision>=0.22.0
transformers>=4.51.3
uvicorn>=0.34.2
uvicorn[standard]>=0.34.2